+++ doc/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 0.2
+* database structure initialization module
+* common libraries module (DB SQL modules)
+* configuration
+
 ## 0.1 - Initial Release
 * project structure - files & folders 
 * docs, license, info etc.